# swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/user/{id}/balance":{"patch":{"operationId":"UserController_chargeBalance","summary":"잔액 충전","description":"잔액을 충전합니다.","parameters":[{"name":"id","required":true,"in":"path","description":"유저 ID","schema":{"example":1,"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateBalanceDto"}}}},"responses":{"200":{"description":"잔액 충전 성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChargeBalanceResponseDto"}}}}},"tags":["User"]},"get":{"operationId":"UserController_getUserBalance","summary":"잔액 조회","description":"현재 잔액을 조회합니다.","parameters":[{"name":"id","required":true,"in":"path","description":"유저 ID","schema":{"example":1,"type":"number"}}],"responses":{"200":{"description":"잔액 조회 성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetBalanceResponseDto"}}}}},"tags":["User"]}},"/concert/{id}/date":{"get":{"operationId":"ConcertController_getAvailableDates","summary":"예약 가능 날짜 조회","description":"특정 콘서트의 예약 가능 날짜를 조회합니다.","parameters":[{"name":"id","required":true,"in":"path","description":"콘서트 ID","schema":{"example":1,"type":"number"}},{"name":"date","required":true,"in":"query","description":"검색 기준 날짜","schema":{"example":"2024-10-17","type":"string"}},{"name":"authorization","required":true,"in":"header","description":"대기열 토큰","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AvailableDatesResponseDto"}}}}},"tags":["Concert"]}},"/concert/{id}/seat":{"get":{"operationId":"ConcertController_getAvailableSeats","summary":"예약 가능한 좌석 조회","parameters":[{"name":"id","required":true,"in":"path","description":"콘서트 ID","schema":{"example":1,"type":"number"}},{"name":"date","required":true,"in":"query","description":"예약 날짜","schema":{"example":"2024-10-17","type":"string"}},{"name":"authorization","required":true,"in":"header","description":"대기열 토큰","schema":{"type":"string"}}],"responses":{"200":{"description":"좌석 조회 성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AvailableSeatsResponseDto"}}}}},"tags":["Concert"]}},"/concert/{id}/reserve-seat":{"post":{"operationId":"ConcertController_reserveSeat","summary":"좌석 임시 예약","description":"좌석을 5분간 임시로 예약합니다.","parameters":[{"name":"id","required":true,"in":"path","description":"콘서트 ID","schema":{"example":1,"type":"number"}},{"name":"authorization","required":true,"in":"header","description":"대기열 토큰","schema":{"type":"string"}},{"name":"date","required":true,"in":"query","description":"예약할 날짜","schema":{"example":"2024-10-17"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReserveSeatRequestDto"}}}},"responses":{"200":{"description":"좌석 임시 예약 성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReserveSeatResponseDto"}}}}},"tags":["Concert"]}},"/payment":{"post":{"operationId":"PaymentController_processPayment","summary":"좌석 결제","description":"예약된 좌석의 결제를 처리합니다.","parameters":[{"name":"authorization","required":true,"in":"header","description":"대기열 토큰","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePaymentDto"}}}},"responses":{"200":{"description":"결제 성공","content":{"application/json":{"schema":{"example":{"status":"Success","tokenState":"Expired","userId":1,"performanceId":1,"seatNumber":25,"amount":10000,"balance":1000}}}}},"402":{"description":"결제 실패","content":{"application/json":{"schema":{"example":{"status":"Fail","userId":1,"performanceId":1,"seatNumber":25,"amount":10000,"balance":1000}}}}}},"tags":["Payment"]}},"/waiting-queue":{"post":{"operationId":"QueueController_issueToken","summary":"토큰 발급","description":"대기열 토큰을 발급합니다.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTokenDto"}}}},"responses":{"200":{"description":"토큰 발급 성공","content":{"application/json":{"schema":{"example":{"tokenState":"WAITING"}}}}}},"tags":["waiting-queue"]},"get":{"operationId":"QueueController_getQueuePosition","summary":"대기열 조회","description":"유저의 대기열 상태 및 순서를 조회합니다.","parameters":[{"name":"authorization","required":true,"in":"header","description":"대기열 토큰","schema":{"type":"string"}}],"responses":{"200":{"description":"대기열 조회 성공","content":{"application/json":{"schema":{"example":{"tokenState":"WAITING"}}}}}},"tags":["waiting-queue"]}}},"info":{"title":"Concert","description":"Concert Reservation Queue API","version":"1.0","contact":{}},"tags":[{"name":"Concert","description":""}],"servers":[],"components":{"schemas":{"UpdateBalanceDto":{"type":"object","properties":{"amount":{"type":"number"}},"required":["amount"]},"ChargeBalanceResponseDto":{"type":"object","properties":{"userId":{"type":"number"},"newBalance":{"type":"number"}},"required":["userId","newBalance"]},"GetBalanceResponseDto":{"type":"object","properties":{"userId":{"type":"number"},"balance":{"type":"number"}},"required":["userId","balance"]},"AvailableDatesResponseDto":{"type":"object","properties":{}},"AvailableSeatsResponseDto":{"type":"object","properties":{}},"ReserveSeatRequestDto":{"type":"object","properties":{}},"ReserveSeatResponseDto":{"type":"object","properties":{}},"CreatePaymentDto":{"type":"object","properties":{"userId":{"type":"number"},"concertId":{"type":"number"},"date":{"type":"string"},"seatNumber":{"type":"number"}},"required":["userId","concertId","date","seatNumber"]},"CreateTokenDto":{"type":"object","properties":{"userId":{"type":"number"}},"required":["userId"]}}}}